word_cloud_generator/requirements.txt
requests
parameterized
python-dateutil
wordcloud
mashumaro